requirements.txt
websockets==12.0
google-generativeai==0.8.3